requirements.txt
amqp==2.5.0
autopep8==1.4.4
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-celery==3.3.0
django-environ==0.4.5
django-redis==4.10.0
django-silk==3.0.2
django-split-settings==0.3.0
djangorestframework==3.10.2
gprof2dot==2017.9.19
idna==2.8
Jinja2==2.10.1
kombu==4.6.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
redis==3.3.4
requests==2.22.0
sentry-sdk==0.10.2
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
